moved everything inside the app directory
