+++ app/package.json
@@ -52,7 +52,6 @@
     "@svgdotjs/svg.js": "^3.0.16",
     "electron-squirrel-startup": "^1.0.0",
     "open": "^7.4.2",
-    "platform-folders": "^0.5.2",
     "ps-list": "^7.2.0",
     "resolve-url-loader": "^3.1.2",
     "winston": "^3.3.3"
