+++ app/src/fileManager/helpers.ts
