+++ app/src/renderer/assets/icons.woff2
