+++ app/src/renderer/clock.ts
