+++ app/src/renderer/isSchedulerRunning.ts
