+++ app/src/renderer/styles.scss
