+++ app/src/runner/index.ts
@@ -1,0 +1,3 @@
+import { ScheduleRunner } from "./scheduleRunner";
+
+export default ScheduleRunner;
