--- fileManager/lib/defaults.d.ts
@@ -1,7 +1,0 @@
-import { Schedule } from "./types";
-export declare const userDataFolder: string;
-export declare const appFolder: string;
-export declare const scheduleFile: string;
-export declare const configFile: string;
-export declare const logFolder: string;
-export declare const defaultSchedule: Schedule;
