--- fileManager/lib/helpers.d.ts
@@ -1,6 +1,0 @@
-import { Task } from "./types";
-export declare function createFolderIfNotExist(folderPath: string): string;
-/**
- * this function opens up the log file for viewing
- */
-export declare function openLogFile(task: Task): void;
