--- fileManager/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "scheduler",
-  "version": "1.0.0",
-  "description": "",
-  "main": "./app/index.js",
-  "scripts": {
-    "test": "env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\" }' mocha -r ts-node/register 'tests/**/*.ts'",
-    "start": "ts-node ./src/index.ts",
-    "start-prod": "node ./build/bundle.js",
-    "build": "tsc",
-    "package": "npm run build && pkg . --out-path ./bin",
-    "pkg": "pkg",
-    "tsc": "tsc",
-    "electron": "electron",
-    "delete-logs": "rm ./logs/error.log ./logs/combined.log"
-  },
-  "bin": "./build/bundle.js",
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/chai": "^4.2.14",
-    "@types/mocha": "^8.2.0",
-    "@types/node": "^14.14.25",
-    "chai": "^4.2.0",
-    "electron": "^11.2.3",
-    "mocha": "^8.2.1",
-    "pkg": "^4.4.9",
-    "ts-loader": "^8.0.17",
-    "ts-node": "^9.1.1",
-    "ts-node-dev": "^1.1.1",
-    "typescript": "^4.1.3",
-    "webpack": "^5.21.2",
-    "webpack-cli": "^4.5.0",
-    "webpack-conditional-loader": "^1.0.12"
-  },
-  "dependencies": {
-    "platform-folders": "^0.5.2",
-    "winston": "^3.3.3"
-  }
-}
