--- runner/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "scheduler-runner",
-  "version": "1.0.0",
-  "description": "",
-  "main": "./app/index.js",
-  "scripts": {
-    "test": "env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\" }' mocha -r ts-node/register 'tests/**/*.ts'",
-    "start": "ts-node ./src/index.ts",
-    "start-prod": "node ./build/bundle.js",
-    "build": "tsc",
-    "package": "npm run build && pkg . --out-path ./bin --targets latest-win-x64",
-    "pkg": "pkg",
-    "tsc": "tsc",
-    "electron": "electron",
-    "delete-logs": "rm ./logs/error.log ./logs/combined.log"
-  },
-  "bin": "./build/runner/src/index.js",
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/chai": "^4.2.14",
-    "@types/mocha": "^8.2.0",
-    "@types/node": "^14.14.25",
-    "chai": "^4.2.0",
-    "mocha": "^8.2.1",
-    "pkg": "^4.4.9",
-    "ts-node": "^9.1.1",
-    "ts-node-dev": "^1.1.1",
-    "typescript": "^4.1.3"
-  },
-  "dependencies": {
-    "open": "^7.4.1",
-    "platform-folders": "^0.5.2",
-    "winston": "^3.3.3"
-  }
-}
