--- runner/src/index.ts
@@ -1,4 +1,0 @@
-import { ScheduleRunner } from "./scheduleRunner";
-
-export const SCHEDULE_RUNNER = new ScheduleRunner();
-export default SCHEDULE_RUNNER;
