--- runner/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "./build"
-  },
-  "extends": "../tsconfig.json",
-  "exclude": ["./tests"],
-  "include": ["./src"]
-}
